rename domain file name, update reference page
